requirements.txt.txt
Flask
flask_sqlalchemy
Flask-Migrate
smtplib
email
PyMySQL